Split out flisp frontend from ast.c and add a plugin interface in between
